requirements.txt
cycler==0.10.0
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
Pillow==6.2.0
pyparsing==2.4.2
python-dateutil==2.8.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
